Tons of refactors
No actual code changes besides deleting some dead code. Moved some
code around and refactored into two new classes, PinHelper and
Commands. These (and the MNA functions) can be further refactored
at a later date.
