+++ src/Common/Block/LightBlock.cs
@@ -18,31 +18,31 @@ using SpiceSharp.Simulations;
 
 namespace ElectricalRevolution
 {
-	public class LightBlock : Block
-	{
-		public override byte[] GetLightHsv(IBlockAccessor blockAccessor, BlockPos pos, ItemStack stack = null)
-		{
-			if (pos != null)
-			{
-        if (blockAccessor.GetBlockEntity(pos) is LightBlockEntity be)
+    public class LightBlock : Block
+    {
+        public override byte[] GetLightHsv(IBlockAccessor blockAccessor, BlockPos pos, ItemStack stack = null)
         {
-          return be.GetLightHsv();
+            if (pos != null)
+            {
+                if (blockAccessor.GetBlockEntity(pos) is LightBlockEntity be)
+                {
+                    return be.GetLightHsv();
+                }
+            }
+            return base.GetLightHsv(blockAccessor, pos, stack);
         }
-      }
-			return base.GetLightHsv(blockAccessor, pos, stack);
-		}
-		public override bool OnBlockInteractStart(IWorldAccessor world, IPlayer byPlayer, BlockSelection blockSel)
-		{
-			BlockEntity blockEntity = world.BlockAccessor.GetBlockEntity(blockSel.Position);
-			LightBlockEntity LblockEntity = (LightBlockEntity)blockEntity;
-			BEBehaviorCreativeConverter behav = blockEntity?.GetBehavior<BEBehaviorCreativeConverter>();
-			if (behav != null)
-			{
-				bool retvalue = behav.OnInteract(byPlayer);
-				//LblockEntity.OnBlockInteract(byPlayer);
-				return retvalue;
-			}
-			return base.OnBlockInteractStart(world, byPlayer, blockSel);
-		}
-	}
+
+        public override bool OnBlockInteractStart(IWorldAccessor world, IPlayer byPlayer, BlockSelection blockSel)
+        {
+            BlockEntity blockEntity = world.BlockAccessor.GetBlockEntity(blockSel.Position);
+            LightBlockEntity LblockEntity = (LightBlockEntity)blockEntity;
+            BEBehaviorCreativeConverter behav = blockEntity?.GetBehavior<BEBehaviorCreativeConverter>();
+            if (behav != null)
+            {
+                bool retvalue = behav.OnInteract(byPlayer);
+                return retvalue;
+            }
+            return base.OnBlockInteractStart(world, byPlayer, blockSel);
+        }
+    }
 }
