+++ src/Common/BlockEntityBehavior/BEBehaviorElectricalConverter.cs
@@ -18,34 +18,39 @@ using SpiceSharp.Simulations;
 
 namespace ElectricalRevolution
 {
-	public class BEBehaviorElectricalConverter : BEBehaviorElectricalNode
-	{
-		public float Powerconverted;
-		public float Efficiency;
-		public BEBehaviorElectricalConverter(BlockEntity blockentity) : base(blockentity){}
-		public override void Initialize(ICoreAPI api, JsonObject properties)
-		{
-			Powerconverted = 0; Efficiency = 1;
-			base.Initialize(api, properties);
-		}
-		public override void GetBlockInfo(IPlayer forPlayer, StringBuilder sb)
-		{
-			base.GetBlockInfo(forPlayer, sb);
-      sb.AppendFormat(Lang.Get("Power Converted: {0}", new object[]
-      {
-        Powerconverted
-      }), Array.Empty<object>()).AppendLine();
-		}
-		public override void FromTreeAttributes(ITreeAttribute tree, IWorldAccessor world)
-		{
-			this.Powerconverted = tree.GetFloat("powerconverted", 0);
-			base.FromTreeAttributes(tree, world);
-		}
-		public override void ToTreeAttributes(ITreeAttribute tree)
-		{
-			tree.SetFloat("powerconverted", this.Powerconverted);
-			base.Voltage = Powerconverted / 100f;
-			base.ToTreeAttributes(tree);
-		}
-	}
+    public class BEBehaviorElectricalConverter : BEBehaviorElectricalNode
+    {
+        public float Powerconverted;
+        public float Efficiency;
+
+        public BEBehaviorElectricalConverter(BlockEntity blockentity) : base(blockentity) {}
+
+        public override void Initialize(ICoreAPI api, JsonObject properties)
+        {
+            Powerconverted = 0; Efficiency = 1;
+            base.Initialize(api, properties);
+        }
+
+        public override void GetBlockInfo(IPlayer forPlayer, StringBuilder sb)
+        {
+            base.GetBlockInfo(forPlayer, sb);
+            sb.AppendFormat(Lang.Get("Power Converted: {0}", new object[]
+            {
+                Powerconverted
+            }), Array.Empty<object>()).AppendLine();
+        }
+
+        public override void FromTreeAttributes(ITreeAttribute tree, IWorldAccessor world)
+        {
+            this.Powerconverted = tree.GetFloat("powerconverted", 0);
+            base.FromTreeAttributes(tree, world);
+        }
+
+        public override void ToTreeAttributes(ITreeAttribute tree)
+        {
+            tree.SetFloat("powerconverted", this.Powerconverted);
+            base.Voltage = Powerconverted / 100f;
+            base.ToTreeAttributes(tree);
+        }
+    }
 }
