+++ src/ElectricalRevolution.cs
@@ -29,22 +29,17 @@ namespace ElectricalRevolution
 		ICoreAPI api = null;
 		ICoreServerAPI sapi = null;
 		ICoreClientAPI capi = null;
-		public Dictionary<BlockPos, BEBehaviorElectricalNode> blockmap = new Dictionary<BlockPos, BEBehaviorElectricalNode>();
-		//warning: These are only instances of blocks, but I made special handling so that you can act on them directly
-		//It's not flawless. When loading from a save file, the block entity will be null until its BE is loaded.
-		//However, any information that was set while the block was unloaded will instantly apply when it is.
 
-		private void OnSaveGameLoading() //Read from loaded save data as the world starts up
-        {
-			//in case of save corruption, uncomment the function below to recover
-			//sapi.WorldManager.SaveGame.StoreData("blockmap",null);
-			byte[] data = sapi.WorldManager.SaveGame.GetData("blockmap");
-            blockmap = data == null ? new Dictionary<BlockPos, BEBehaviorElectricalNode>() : SerializerUtil.Deserialize<Dictionary<BlockPos, BEBehaviorElectricalNode>>(data);
-        }
-		private void OnSaveGameSaving() //Write save data to the world as it shuts down
-        {
-            sapi.WorldManager.SaveGame.StoreData("blockmap", SerializerUtil.Serialize(blockmap));
-        }
+        ///warning: These are only instances of blocks, but I made special handling so that you can act on them directly
+        ///It's not flawless. When loading from a save file, the block entity will be null until its BE is loaded.
+        ///However, any information that was set while the block was unloaded will instantly apply when it is.
+		public Dictionary<BlockPos, BEBehaviorElectricalNode> blockmap = new Dictionary<BlockPos, BEBehaviorElectricalNode>();
+
+        public Dictionary<string,SpiceSharp.Entities.IEntity> componentlist = new Dictionary<string,SpiceSharp.Entities.IEntity>();
+        public Dictionary<string,SpiceSharp.Simulations.TimeParameters> TPlist = new Dictionary<string,SpiceSharp.Simulations.TimeParameters>();
+
+        public BlockPos[] leadermap = new BlockPos[0];
+
 		public override void Start(ICoreAPI api)
 		{
 			this.api = api;
@@ -60,7 +55,36 @@ namespace ElectricalRevolution
 			api.RegisterBlockEntityBehaviorClass("MPGenerator", typeof(BEBehaviorMPGenerator));
 			api.RegisterBlockEntityBehaviorClass("CreativeConverter",typeof(BEBehaviorCreativeConverter));
 		}
-		public override void StartClientSide(ICoreClientAPI capi){this.capi = capi;}
+
+        public override void StartClientSide(ICoreClientAPI capi)
+        {
+            this.capi = capi;
+        }
+
+        public override void StartServerSide(ICoreServerAPI sapi)
+        {
+            this.sapi = sapi;
+            sapi.Event.SaveGameLoaded += OnSaveGameLoading;
+            sapi.Event.GameWorldSave += OnSaveGameSaving;
+            sapi.World.RegisterGameTickListener(TickMNA, 1000); //tick the MNA every second
+            Commands.registerCommands(api, sapi, this);
+        }
+
+        /// Read from loaded save data as the world starts up
+        private void OnSaveGameLoading()
+        {
+            //in case of save corruption, uncomment the function below to recover
+            //sapi.WorldManager.SaveGame.StoreData("blockmap",null);
+            byte[] data = sapi.WorldManager.SaveGame.GetData("blockmap");
+            blockmap = data == null ? new Dictionary<BlockPos, BEBehaviorElectricalNode>() : SerializerUtil.Deserialize<Dictionary<BlockPos, BEBehaviorElectricalNode>>(data);
+        }
+
+        ///Write save data to the world as it shuts down
+        private void OnSaveGameSaving()
+        {
+            sapi.WorldManager.SaveGame.StoreData("blockmap", SerializerUtil.Serialize(blockmap));
+        }
+
 		private static IEnumerable<double> TimePoints
         {
             get
@@ -73,11 +97,10 @@ namespace ElectricalRevolution
                 }    
             }
         }
-		public Dictionary<string,SpiceSharp.Entities.IEntity> componentlist = new Dictionary<string,SpiceSharp.Entities.IEntity>();
-		public Dictionary<string,SpiceSharp.Simulations.TimeParameters> TPlist = new Dictionary<string,SpiceSharp.Simulations.TimeParameters>();
 
+        ///Tries to find neighbours and attach them to eachother
 		public void UpdateBlockmap(BlockPos thispos)
-		{//Tries to find neighbours and attach them to eachother
+		{
 			BlockPos[] neighbourposes = new BlockPos[6]{thispos.UpCopy(1),thispos.DownCopy(1),thispos.NorthCopy(1),thispos.EastCopy(1),thispos.SouthCopy(1),thispos.WestCopy(1),};
 			foreach(BlockPos thatpos in neighbourposes)
 			{
@@ -153,10 +176,11 @@ namespace ElectricalRevolution
 			}
 			CreateLeaderMap();
 		}
-		public BlockPos[] leadermap = new BlockPos[0];
-		public void CreateLeaderMap()//this is what compiles the blockmap into many seperate leadermaps so they can be turned into ckts and ultimately trans
-		{//unlike the leadermap, it's not saved since it's created from the blockmap every tick
-			
+
+        ///this is what compiles the blockmap into many seperate leadermaps so they can be turned into ckts and ultimately trans
+        ///unlike the blockmap, the leadermap is not saved, since it's created from the blockmap every tick
+		public void CreateLeaderMap()
+		{
 			leadermap = new BlockPos[0]; //first we wipe it clean to prevent any pollution from last tick
 			foreach(KeyValuePair<BlockPos,BEBehaviorElectricalNode> entry in blockmap)
 			{
@@ -165,24 +189,24 @@ namespace ElectricalRevolution
 					leadermap = leadermap.Append(entry.Key); //add it to the leadermap
 				}
 			}
-
-			
 		}
+
+        ///tiebreaker function. Smallest X, then Y, then Z wins. True if thisnode wins. False if thatnode wins.
+        /// TODO: Replace with a BlockPos comparison function extension (if that exists in C#) so you can do `poslocal < posremote`
 		public bool NodeTieBreaker(BlockPos poslocal, BlockPos posremote)
-    	{ //tiebreaker function. Smallest X, then Y, then Z wins. True if thisnode wins. False if thatnode wins.
+    	{
       		if(poslocal.X < posremote.X){return true;}
       		if(poslocal.X > posremote.X){return false;}
 			//in most cases, it will resolve here, but in the case of a tie...
 			if(poslocal.Y < posremote.Y){return true;}
      	 	if(poslocal.Y > posremote.Y){return false;}
-    		//gee golly, still not resolved? 
+    		  //gee golly, still not resolved?
       		if(poslocal.Z < posremote.Z){return true;}
     	 	if(poslocal.Z > posremote.Z){return false;}
     	 	throw new ArgumentException//something clearly went wrong. Throw an exception.
       		("ShouldIBeTheLeader was unable to resolve. It's likely because the local and remote position were the same.");
    		}
-		
-		
+
 		public void CreateCircuit(Transient tran, out Circuit ckt)
 		{
 			ckt = new Circuit();
@@ -195,21 +219,23 @@ namespace ElectricalRevolution
 			{
 				string componentname = entry.Key;
 				SpiceSharp.Entities.IEntity component = entry.Value;
-				string componenttype = GetNodeTypeFromName(componentname);
-				if(ckt.TryGetEntity(componentname, out var throwaway)){continue;}//only add to the ckt if it wasn't there before
+                string componenttype = PinHelper.GetNodeTypeFromName(componentname);
+				if(ckt.TryGetEntity(componentname, out var throwaway)){continue;} //only add to the ckt if it wasn't there before
 				AddComponent(tran,ckt,componentname,component);
 			}
 		}
+
 		public void OnMNAExport(object sender, ExportDataEventArgs exargs)
 		{
-			Transient tran = (Transient)sender;
-
-			if(tran.TimeParameters.StopTime <= 0)
-			{
-			MNAFinish(tran); //Tell the MNA to upload the information from this tick into the next so that it can pass on.
-			}
+            Transient tran = (Transient)sender;
+            if(tran.TimeParameters.StopTime <= 0)
+            {
+                MNAFinish(tran); //Tell the MNA to upload the information from this tick into the next so that it can pass on.
+            }
 		}
-		public void MNAFinish(Transient tran) //finish the tick then shut down
+
+        ///finish the tick then shut down
+		public void MNAFinish(Transient tran)
 		{
 			//foreach(KeyValuePair<string,Export<IBiasingSimulation, double>> entry in ICWatchers)
 			//{
@@ -223,21 +249,22 @@ namespace ElectricalRevolution
 			//}
 			castMNAtoblocks(tran); //sends data from the MNA into the blockents
 		}
+
 		public void castMNAtoblocks(Transient tran)
 		{
 			foreach(KeyValuePair<string,SpiceSharp.Entities.IEntity> entry in componentlist)
 			{
 				SpiceSharp.Entities.IEntity component = entry.Value;
 				string componentname = component.Name;
-				string componentlocation = GetPositivePin(componentname);
-				BlockPos blockpos = PinToBlockPos(componentlocation,out Vec3i sublocation);
+				string componentlocation = PinHelper.GetPositivePin(componentname);
+                BlockPos blockpos = PinHelper.PinToBlockPos(componentlocation,out Vec3i sublocation);
 				//Block block = api.World.BlockAccessor.GetBlock(blockpos);
 				BlockEntity blockent = api.World.BlockAccessor.GetBlockEntity(blockpos);
 				if(blockent == null){continue;} //it's probably unloaded. No need to update it.
 				BEBehaviorElectricalNode blockbehavior = blockent.GetBehavior<BEBehaviorElectricalNode>();
 				if(blockbehavior == null){continue;} //if it's not a node, don't bother with it.
 				blockbehavior.Current = new RealCurrentExport(tran,componentname).Value;
-				blockbehavior.Voltage = new RealVoltageExport(tran,GetPositivePin(componentname),GetNegativePin(componentname)).Value;
+                blockbehavior.Voltage = new RealVoltageExport(tran,PinHelper.GetPositivePin(componentname),PinHelper.GetNegativePin(componentname)).Value;
 				
 				//sapi.BroadcastMessageToAllGroups(blockbehavior.Voltage + "+"+ blockbehavior.Current,EnumChatType.CommandError);
 				blockbehavior.Blockentity.MarkDirty(true);
@@ -245,99 +272,7 @@ namespace ElectricalRevolution
 
 			}
 		}
-		public override void StartServerSide(ICoreServerAPI sapi)
-		{
-			this.sapi = sapi;
 
-			sapi.Event.SaveGameLoaded += OnSaveGameLoading;
-            sapi.Event.GameWorldSave += OnSaveGameSaving;
-			
-			/*string voltagename = GetNodeNameFromPins("VoltageSource",GetPinNameAtPosition(new BlockPos(10,3,10),new Vec3i(0,0,0)),"0");
-			string resistorname = GetNodeNameFromPins("Resistor",GetPinNameAtPosition(new BlockPos(10,3,10),new Vec3i(0,0,0)),GetPinNameAtPosition(new BlockPos(11,3,10),new Vec3i(0,0,0)));
-			string diodename = GetNodeNameFromPins("Diode",GetPinNameAtPosition(new BlockPos(11,3,10),new Vec3i(0,0,0)),GetPinNameAtPosition(new BlockPos(12,3,10),new Vec3i(0,0,0)));
-			DiodeModel diodemodel = CreateDiodeModel(diodename);
-			string diodemodelname = diodemodel.Name;
-			string inductorname = GetNodeNameFromPins("Inductor",GetPinNameAtPosition(new BlockPos(12,3,10),new Vec3i(0,0,0)),GetPinNameAtPosition(new BlockPos(13,3,10),new Vec3i(0,0,0)));
-			string capacitorname = GetNodeNameFromPins("Capacitor",GetPinNameAtPosition(new BlockPos(13,3,10),new Vec3i(0,0,0)),"0");
-			
-			//make some basic components with hard-coded locations, you know, for testing.
-			componentlist.Add(voltagename, new VoltageSource(voltagename,GetPositivePin(voltagename),GetNegativePin(voltagename),1));
-			componentlist.Add(resistorname, new Resistor(resistorname,GetPositivePin(resistorname),GetNegativePin(resistorname),1));
-			componentlist.Add(diodename, new Diode(diodename,GetPositivePin(diodename),GetNegativePin(diodename),diodemodelname));
-			componentlist.Add(inductorname, new Inductor(inductorname,GetPositivePin(inductorname),GetNegativePin(inductorname),1));
-			componentlist.Add(capacitorname, new Capacitor(capacitorname,GetPositivePin(capacitorname),GetNegativePin(capacitorname),1)); */
-
-			sapi.World.RegisterGameTickListener(TickMNA,1000); //tick the MNA every second
-
-			sapi.RegisterCommand("mna","Gets a readout of the leaderlist","",(IServerPlayer splayer, int groupId, CmdArgs args) =>
-            {
-				string message = "";
-				foreach(BlockPos leaderpos in leadermap)
-				{
-					message = message + leaderpos + "[ ";
-					BlockPos[] recruiterlist = blockmap[leaderpos].NodeList;
-					foreach(BlockPos recruiterpos in recruiterlist)
-					{
-						message = message + recruiterpos + "  ";
-					}
-					message = message + "]";
-				}
-				sapi.SendMessage(splayer,GlobalConstants.GeneralChatGroup,message,EnumChatType.CommandSuccess);
-			/*if(args.Length > 0){
-			Double.TryParse(args[0],out double dcsetting);
-			ckt.TryGetEntity(GetNodeNameFromPins("VoltageSource",GetPinNameAtPosition(new BlockPos(10,3,10),new Vec3i(0,0,0)),"0"), out SpiceSharp.Entities.IEntity component);
-			if(!component.TrySetParameter("dc",dcsetting)){throw new NullReferenceException("MNA command couldn't set the parameter: dc");}
-			}
-
-			string capacitorname = GetNodeNameFromPins("Capacitor",GetPinNameAtPosition(new BlockPos(13,3,10),new Vec3i(0,0,0)),"0");
-			string inductorname = GetNodeNameFromPins("Inductor",GetPinNameAtPosition(new BlockPos(12,3,10),new Vec3i(0,0,0)),GetPinNameAtPosition(new BlockPos(13,3,10),new Vec3i(0,0,0)));
-			//double inductorreadout = ICWatchers[inductorname].Value;
-			//double capacitorreadout = ICWatchers[capacitorname].Value;
-			ckt.TryGetEntity(capacitorname, out SpiceSharp.Entities.IEntity node);
-			if(!node.TryGetProperty("ic", out double nodeic)){throw new NullReferenceException("MNA command couldn't get the property: ic");}
-			double capacitorreadout = nodeic;
-			ckt.TryGetEntity(inductorname, out node);
-			if(!node.TryGetProperty("ic", out nodeic)){throw new NullReferenceException("MNA command couldn't get the property: ic");}
-			double inductorreadout = nodeic;
-			sapi.BroadcastMessageToAllGroups("capacitor: "+capacitorreadout+" inductor: "+inductorreadout,EnumChatType.CommandSuccess);
-			*/
-			}, Privilege.chat);
-
-			sapi.RegisterCommand("blocklist","Read out the block list","",(IServerPlayer splayer, int groupId, CmdArgs args) =>
-            {
-				double inputvalue = 0;
-				if(args.Length > 0){
-				Double.TryParse(args[0],out inputvalue);
-				}
-				foreach(KeyValuePair<BlockPos,BEBehaviorElectricalNode> entry in blockmap)
-				{
-					string message = "Block at: " + entry.Key;
-					message = message + " Internal hash: " + entry.Value.Blockentity?.GetHashCode().ToString();
-					message = message + " External hash: " + api.World.BlockAccessor.GetBlockEntity(entry.Value.Blockentity.Pos)?.GetHashCode().ToString();
-					message = message + " Resistance: " + entry.Value.Resistance;
-					entry.Value.Resistance = inputvalue;
-					sapi.SendMessage(splayer,GlobalConstants.GeneralChatGroup,message,EnumChatType.CommandSuccess);
-					entry.Value.Blockentity.MarkDirty(true);
-				}
-			}, Privilege.chat);
-
-			/*sapi.RegisterCommand("here","Where am I? answered in text form","",(IServerPlayer splayer, int groupId, CmdArgs args) =>
-            {
-				BlockPos blockpos = splayer.Entity.Pos.AsBlockPos;
-				Vec3i subblockpos = new Vec3i(0,0,0);
-
-				string message = GetPinNameAtPosition(blockpos,subblockpos);
-				string nodename = GetNodeNameFromPins("VoltageSource",GetPinNameAtPosition(blockpos,subblockpos),"0");
-				string nodetype = GetNodeTypeFromName(nodename);
-				string pospinname = GetPositivePin(nodename);
-				string negpinname = GetNegativePin(nodename);
-				splayer.SendMessage(GlobalConstants.GeneralChatGroup,message,EnumChatType.CommandSuccess);
-				splayer.SendMessage(GlobalConstants.GeneralChatGroup,nodename,EnumChatType.CommandSuccess);
-				splayer.SendMessage(GlobalConstants.GeneralChatGroup,nodetype,EnumChatType.CommandSuccess);
-				splayer.SendMessage(GlobalConstants.GeneralChatGroup,pospinname,EnumChatType.CommandSuccess);
-				splayer.SendMessage(GlobalConstants.GeneralChatGroup,negpinname,EnumChatType.CommandSuccess);
-			}, Privilege.chat);*/
-		}
 		public void TickMNA(float par)
 		{
 			if(blockmap.Count < 1){return;}//list is null. Don't start yet.
@@ -366,6 +301,7 @@ namespace ElectricalRevolution
 				tran.Run(ckt);
 			}
 		}
+
 		public DiodeModel CreateDiodeModel(string name)
         {
             var diodemodel = new DiodeModel(name + "#diodemodel"); //the parameters of a standard silicon diode
@@ -377,152 +313,65 @@ namespace ElectricalRevolution
 			diodemodel.SetParameter("tt",20e-9); //transit time
             return diodemodel;
         }
-		public static string GetPinNameAtPosition(BlockPos blockpos, Vec3i subblockpos) //Converts blockpos and subblockpos into a pin
-		{
-			if(blockpos == null || subblockpos == null){return "0";}
-			string returnstring = "" + blockpos + " (" + subblockpos.X + ", " + subblockpos.Y + ", "+ subblockpos.Z + ")";
-			return returnstring;
-		}
-		public static string GetNodeNameFromPins(string componenttype, string pinnamepos, string pinnameneg) //Converts a pinpos and a pinneg into a nodename
-		{
-			string returnstring = componenttype + ":" + pinnamepos + "~" + pinnameneg;
-			return returnstring;
-		}
-		public static string GetNodeTypeFromName(string nodename)
-		{
-			string returnstring = nodename;
-			int stopat = returnstring.IndexOf(":");
-			return returnstring.Substring(0, stopat);
-		}
-		public static string GetPositivePin(string nodename) //Extracts the positive pin from a node's name.
-		{
-			string returnstring = nodename;
-			int startat = returnstring.IndexOf(":") + 1;
-			int stopat = returnstring.IndexOf("~");
-			int length = stopat-startat;
-			return returnstring.Substring(startat, length);
-		}
-		public static string GetNegativePin(string nodename) //Extracts the negative pin from the node's name.
-		{
-			string returnstring = nodename;
-			int startat = returnstring.IndexOf("~") + 1;
-			return returnstring.Substring(startat);
-		}
-		public static BlockPos PinToBlockPos(string pinname,out Vec3i sublocation) //converts a pin into a blockpos
-		{
-			string workstring = pinname;
-			//the pin name will look like this:
-			//10, 3, 10 (0, 0, 0)
-			int startat = 0;
-			int stopat = workstring.IndexOf(",");
-			int length = stopat-startat;
-			int x = workstring.Substring(startat,length).ToInt(-666);
-			workstring = workstring.Substring(stopat+2); //2 gets rid of the comma and space
-			//3, 10 (0, 0, 0)
-			stopat = workstring.IndexOf(",");
-			length = stopat-startat;
-			int y = workstring.Substring(startat,length).ToInt(-666);
-			workstring = workstring.Substring(stopat+2); //2 gets rid of the comma and space
-			//should be 10 (0, 0, 0)
-			stopat = workstring.IndexOf("(")-1;
-			length = stopat-startat;
-			int z = workstring.Substring(startat,length).ToInt(-666);
-			workstring = workstring.Substring(stopat+2); //get rid of the space and )
-			//0, 0, 0)
-			stopat = workstring.IndexOf(",");
-			length = stopat-startat;
-			int sex = workstring.Substring(startat,length).ToInt(-666); //sub-x. What the E stands for is as much of a mystery as what the N in ELN stands for.
-			workstring = workstring.Substring(stopat+2);
-			//should be 0, 0)
-			stopat = workstring.IndexOf(",");
-			length = stopat-startat;
-			int sey = workstring.Substring(startat,length).ToInt(-666);
-			workstring = workstring.Substring(stopat+2);
-			//should be 0)
-			stopat = workstring.IndexOf(")");
-			length = stopat-startat;
-			int sez = workstring.Substring(startat,length).ToInt(-666);
-			workstring = workstring.Substring(stopat+1); //should be empty now
 
-			sublocation = new Vec3i(sex,sey,sez);
-			//sapi.BroadcastMessageToAllGroups("x:"+x+"y:"+y+"z:"+z+ ": "+workstring + " subblock:" + sublocation.ToString(),EnumChatType.CommandError);
-			//x:11y:3z:10:  subblock:X=1,Y=0,Z=0
-			
-			return new BlockPos(x,y,z);
-		}
 		public void AddComponent(Transient tran, Circuit ckt,string nodename,SpiceSharp.Entities.IEntity component)
-		{
+        {
+            //in (paracap) -> pararesistor -> component+ -> component- -> parainduct -> out
+            string pos = PinHelper.GetPositivePin(nodename);
+            string neg = PinHelper.GetNegativePin(nodename);
+            /*Circuit subcircuit = new Circuit(
+            //new Resistor (nodename+"PRP",pos,subpos,0.01),
+            //new Capacitor(nodename+"PCP",subpos,"0",0.01),
+            //new Inductor (nodename+"PI",subneg,neg,0.01),
+            //new Capacitor(nodename+"PCN",neg,"0",0.01)
+            );*/
+            string componenttype = PinHelper.GetNodeTypeFromName(nodename);
+            switch(componenttype)
+            {
+                case "VoltageSource":
+                    ckt.Add(component as VoltageSource);
+                    break;
 
-			//in (paracap) -> pararesistor -> component+ -> component- -> parainduct -> out
+                case "Resistor":
+                    ckt.Add(component as Resistor);
+                    break;
 
-			string pos = GetPositivePin(nodename);
-			string neg = GetNegativePin(nodename);
-			/*Circuit subcircuit = new Circuit(
-			//new Resistor (nodename+"PRP",pos,subpos,0.01),
-			//new Capacitor(nodename+"PCP",subpos,"0",0.01),
-			//new Inductor (nodename+"PI",subneg,neg,0.01),
-			//new Capacitor(nodename+"PCN",neg,"0",0.01)
-			);*/
-			string componenttype = GetNodeTypeFromName(nodename);
-			switch(componenttype)
-				{
-					case "VoltageSource":
-					ckt.Add(component as VoltageSource);
-					break;
+                case "Inductor":
+                    ckt.Add(component as Inductor);
+                    break;
 
-					case "Resistor":
-					ckt.Add(component as Resistor);
-					break;
+                case "Capacitor":
+                    ckt.Add(component as Capacitor);
+                    break;
 
-					case "Inductor":
-					ckt.Add(component as Inductor);
-					break;
+                case "Diode":
+                    ckt.Add(component as Diode);
+                    ckt.Add(CreateDiodeModel(component.Name));
+                    break;
 
-					case "Capacitor":
-					ckt.Add(component as Capacitor);
-					break;
+                default:
+                    break;
+            }
+            //now we check the pos and neg pins to make sure they have unideal components
+            //they need to be there to ensure all components are self-sufficient and won't crash
 
-					case "Diode":
-					ckt.Add(component as Diode);
-					ckt.Add(CreateDiodeModel(component.Name));
-					break;
-
-					default:
-					break;
-				}
-				//now we check the pos and neg pins to make sure they have unideal components
-				//they need to be there to ensure all components are self-sufficient and won't crash
-			
-			
-			
-			if(!pos.EqualsFast("0"))//prevents the creation of unideal components on earth ground, as they're not needed there
-			{
-			string subpos = pos + "UNIDEAL";
-			if(!ckt.TryGetEntity(GetNodeNameFromPins("Resistor",pos,subpos),out var throwaway)) //no need to add the component if it already exists
-			{
-				ckt.Add(new Resistor(GetNodeNameFromPins("Resistor",pos,subpos),pos,subpos,0.01)); //add a resistor with 0.01ohm
-				ckt.Add(new Capacitor(GetNodeNameFromPins("Capacitor",subpos,"0"),subpos,"0",0.01)); //add a capacitor-To-Ground with 0.01 farads
-			} 
-			}
-			if(!neg.EqualsFast("0")){
-			string subneg = neg + "UNIDEAL";
-			if(!ckt.TryGetEntity(GetNodeNameFromPins("Resistor",neg,subneg),out var throwaway)) //no need to add the component if it already exists
-			{
-				ckt.Add(new Resistor(GetNodeNameFromPins("Resistor",neg,subneg),neg,subneg,0.01)); //add a resistor with 0.01ohm
-				ckt.Add(new Capacitor(GetNodeNameFromPins("Capacitor",subneg,"0"),subneg,"0",0.01)); //add a capacitor-To-Ground with 0.01 farads
-			}
-			}
-
-
-		}
-		public string Addifnot(string inputstring, string addthis, string ifnotthis)
-		{
-			string outputstring = inputstring;
-			if(!inputstring.EqualsFast(ifnotthis))
-			{
-				inputstring = inputstring + addthis;
-			}
-			return outputstring;
-		}
+            if(!pos.EqualsFast("0"))//prevents the creation of unideal components on earth ground, as they're not needed there
+            {
+                string subpos = pos + "UNIDEAL";
+                if(!ckt.TryGetEntity(PinHelper.GetNodeNameFromPins("Resistor",pos,subpos),out var throwaway)) //no need to add the component if it already exists
+                {
+                    ckt.Add(new Resistor(PinHelper.GetNodeNameFromPins("Resistor",pos,subpos),pos,subpos,0.01)); //add a resistor with 0.01ohm
+                    ckt.Add(new Capacitor(PinHelper.GetNodeNameFromPins("Capacitor",subpos,"0"),subpos,"0",0.01)); //add a capacitor-To-Ground with 0.01 farads
+                }
+            }
+            if(!neg.EqualsFast("0")){
+                string subneg = neg + "UNIDEAL";
+                if(!ckt.TryGetEntity(PinHelper.GetNodeNameFromPins("Resistor",neg,subneg),out var throwaway)) //no need to add the component if it already exists
+                {
+                    ckt.Add(new Resistor(PinHelper.GetNodeNameFromPins("Resistor",neg,subneg),neg,subneg,0.01)); //add a resistor with 0.01ohm
+                    ckt.Add(new Capacitor(PinHelper.GetNodeNameFromPins("Capacitor",subneg,"0"),subneg,"0",0.01)); //add a capacitor-To-Ground with 0.01 farads
+                }
+            }
+        }
 	}
 }
